Refactored app structure
 - Created components/wrappers
 - Created public images
 - Added know URIs to Constants.js
 - New FusedDeck component
 - Created containers folder
 - Made DeckTable more agnostic, to be able to reuse for fused decks
