+++ src/App.js
@@ -1,18 +1,20 @@
-import { Box, Button, CssBaseline, List, ListItem, ListItemText, Typography } from '@mui/material';
+import { Box, CssBaseline, List, ListItem, ListItemText, Typography } from '@mui/material';
 import * as React from 'react';
 import Search from './components/Search';
-import DeckTable from './components/DeckTable'
 import DeckLink from './components/DeckLink';
-import ButtonWrapper from './components/ButtonWrapper';
-import BoxStyles from './components/BoxWrapper';
+import ButtonWrapper from './components/wrappers/ButtonWrapper';
+import DeckTableContainer from './containers/DeckTableContainer';
+import {URIs} from "./Constants";
+import FusedDeck from "./components/FusedDeck";
 
 export function App() {
+
   //state
   const [decks, setDecks] = React.useState([])
   const [fused, setFused] = React.useState([])
 
   function handleSearchClick(username) {
-    fetch('https://ul51g2rg42.execute-api.us-east-1.amazonaws.com/main/deck/?pageSize=1000&inclCards=true&username=' + username)
+    fetch(URIs.SolforgeAPI + `deck?pageSize=1000&inclCards=true&username=${username}`)
       .then(response => response.json())
       .then(data => setDecks(data.Items))
   }
@@ -28,35 +30,14 @@ export function App() {
     }
   }
 
-  function showFused() {
-    return (
-      <Box sx={{margin: 2, border: 2, width: "16%", borderColor: 'grey.500', bgcolor: 'background.paper', borderRadius: 6}} justifyContent="center">
-        <List>
-          <ListItem>
-            <ListItemText>
-              <Typography variant="h6">Your random fused deck:</Typography>
-            </ListItemText>
-          </ListItem>
-          {fused.map(deck => {
-            return (
-              <ListItem key={deck.id}>
-                <ListItemText id={deck.id} primary={<DeckLink deckId={deck.id} deckName={deck.name} />} secondary={deck.faction} />
-              </ListItem>
-            )
-          })}
-        </List>
-      </Box>
-    )
-  }
 
   return (
     <>
       <CssBaseline />
       <Search fieldLabel="Username" buttonLabel="Import" handleClick={handleSearchClick} />
-      {decks.length > 0 && <DeckTable decks={decks} />}
-      {fused.length > 0 && showFused()}
+      {fused.length > 0 && <FusedDeck fused={fused} />}
       {decks.length > 0 && <ButtonWrapper handleClick={randomFused}>Random Fused</ButtonWrapper>}
-
+      {decks.length > 0 && <DeckTableContainer decks={decks} />}
     </>
   )
 }
