+++ src/Constants.js
@@ -1,0 +1,13 @@
+const IMAGE_ROOT = "images"
+export const ImagePaths = {
+    Level1: `${IMAGE_ROOT}/level1.svg`,
+    Level2: `${IMAGE_ROOT}/level2.svg`,
+    Level3: `${IMAGE_ROOT}/level3.svg`,
+    Attack: `${IMAGE_ROOT}/attack.svg`,
+    Health: `${IMAGE_ROOT}/health.svg`
+}
+
+export const URIs = {
+    SolforgeSite: "https://solforgefusion.com/",
+    SolforgeAPI: "https://ul51g2rg42.execute-api.us-east-1.amazonaws.com/main/"
+}
