--- src/components/Constants.js
@@ -1,7 +1,0 @@
-export const ImagePaths = {
-    Level1: "./Level1.svg",
-    Level2: "./Level2.svg",
-    Level3: "./Level3.svg",
-    Attack: "./attack.svg",
-    Health: "./health.svg"
-}
