+++ src/components/DeckLink.js
@@ -1,5 +1,6 @@
 import { Link } from '@material-ui/core'
 import * as React from 'react'
+import {URIs} from "../Constants";
 
 
 
@@ -9,7 +10,7 @@ export default function DeckLink(props) {
   const { deckId, deckName } = props
 
   return (
-    <Link variant="subtitle2" href={"https://solforgefusion.com/decks/" + deckId} target="_blank">{deckName}</Link>
+    <Link variant="subtitle2" href={ URIs.SolforgeSite + "decks/" + deckId} target="_blank">{deckName}</Link>
   )
 }
