+++ src/components/DeckTable.js
@@ -1,10 +1,7 @@
-import { Box, makeStyles, Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, TableSortLabel } from '@material-ui/core';
+import { makeStyles, Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, TableSortLabel } from '@material-ui/core';
 import * as React from 'react'
-import DeckLink from './DeckLink'
-import { visuallyHidden } from '@mui/utils'
-import IconWrapper from './IconWrapper';
-import { Typography } from '@mui/material';
-import { ImagePaths } from './Constants';
+import DeckLink from "./DeckLink";
+
 
 
 const useStyles = makeStyles({
@@ -13,103 +10,42 @@ const useStyles = makeStyles({
   },
 });
 
+/**
+ *
+ * @param {*} props
+ * props.headers
+ * @returns
+ */
 export default function DeckTable(props) {
   const classes = useStyles();
 
   //props
-  const { decks } = props
+  const { headers, rows, tableName } = props
 
   //state
   const [order, setOrder] = React.useState('asc');
   const [orderBy, setOrderBy] = React.useState('name');
-
-  const CardType = {
-    Creature: "Creature",
-    Spell: "Spell"
-  }
-
-  function headerIcons(label, imagePaths) {
-    return <><IconWrapper path={imagePaths}></IconWrapper> <Typography variant="inherit" sx={{ p: 0.2 }}>{label}</Typography></>
-  }
-  const headers = [
-    { id: "name", label: "Name", number: false },
-    { id: "faction", label: "Faction", number: false },
-    { id: "attackI", label: headerIcons("Average Attack", ImagePaths.Level1), number: true },
-    { id: "attackII", label: headerIcons("Average Attack", ImagePaths.Level2), number: true },
-    { id: "attackIII", label: headerIcons("Average Attack", ImagePaths.Level3), number: true },
-    { id: "healthI", label: headerIcons("Average Health", ImagePaths.Level1), number: true },
-    { id: "healthII", label: headerIcons("Average Health", ImagePaths.Level2), number: true },
-    { id: "healthIII", label: headerIcons("Average Health", ImagePaths.Level3), number: true },
-    { id: "nbCreatures", label: "Creatures", number: true },
-    { id: "nbSpells", label: "Spells", number: true }
-  ]
-
-  function getRows() {
-    return decks.map(deck => {
-      let deckTotals = Object.values(deck.cards).reduce((totals, card) => {
-        if (card.cardType === CardType.Creature) {
-          totals.nbCreatures += 1
-          totals.totalAttackI += card.levels["1"].attack
-          totals.totalAttackII += card.levels["2"].attack
-          totals.totalAttackIII += card.levels["3"].attack
-          totals.totalHealthI += card.levels["1"].health
-          totals.totalHealthII += card.levels["2"].health
-          totals.totalHealthIII += card.levels["3"].health
-        } else if (card.cardType === CardType.Spell) {
-          totals.nbSpells += 1
-        }
-        return totals
-      },
-        {
-          totalAttackI: 0,
-          totalAttackII: 0,
-          totalAttackIII: 0,
-          totalHealthI: 0,
-          totalHealthII: 0,
-          totalHealthIII: 0,
-          nbCreatures: 0,
-          nbSpells: 0
-        })
-      return {
-        "id": deck.id,
-        "name": deck.name,
-        "faction": deck.faction,
-        "attackI": getFormattedAverage(deckTotals.totalAttackI, deckTotals.nbCreatures),
-        "attackII": getFormattedAverage(deckTotals.totalAttackII, deckTotals.nbCreatures),
-        "attackIII": getFormattedAverage(deckTotals.totalAttackIII, deckTotals.nbCreatures),
-        "healthI": getFormattedAverage(deckTotals.totalHealthI, deckTotals.nbCreatures),
-        "healthII": getFormattedAverage(deckTotals.totalHealthII, deckTotals.nbCreatures),
-        "healthIII": getFormattedAverage(deckTotals.totalHealthIII, deckTotals.nbCreatures),
-        "nbCreatures": deckTotals.nbCreatures,
-        "nbSpells": deckTotals.nbSpells
-      }
-    })
-  }
-
-  function getFormattedAverage(total, count, decimals = 3) {
-    return (total / count).toFixed(decimals)
-  }
-
-  function getTableRows() {
-    const rows = getRows()
-    return rows.sort(getComparator(order, orderBy)).map(row => {
-      return (
-        <TableRow key={row.id}>
-          {headers.map(header => (
-            <TableCell align={header.number ? "right" : "left"}>
-              {header.id === 'name' ? <DeckLink deckId={row.id} deckName={row[header.id]} /> : row[header.id]}
-            </TableCell>
-          ))}
-        </TableRow>
-      )
-    })
-  }
 
   function createSortHandler(event, property) {
     const isAsc = orderBy === property && order === 'asc';
     setOrder(isAsc ? 'desc' : 'asc');
     setOrderBy(property);
   }
+
+  function getSortedTableRows() {
+    return rows.sort(getComparator(order, orderBy)).map(row => {
+      return (
+          <TableRow key={row.id}>
+            {headers.map(header => (
+                <TableCell align={header.number ? "right" : "left"}>
+                  {header.id === 'name' ? <DeckLink deckId={row.id} deckName={row[header.id]} /> : row[header.id]}
+                </TableCell>
+            ))}
+          </TableRow>
+      )
+    })
+  }
+
 
   function getComparator(order, orderBy) {
     const isNumber = headers.find(header => header.id === orderBy).number
@@ -133,7 +69,7 @@ export default function DeckTable(props) {
 
   return (
     <TableContainer component={Paper}>
-      <Table className={classes.table} aria-label="simple table">
+      <Table className={classes.table} aria-label={tableName}>
         <TableHead>
           <TableRow>
             {headers.map(header => (
@@ -150,7 +86,7 @@ export default function DeckTable(props) {
           </TableRow>
         </TableHead>
         <TableBody>
-          {getTableRows()}
+          {getSortedTableRows()}
         </TableBody>
       </Table>
     </TableContainer>
